+++ argilla-server/src/argilla_server/apis/v1/handlers/documents.py
@@ -5,7 +5,7 @@ from uuid import UUID
 from typing import TYPE_CHECKING, Optional, List, Union
 
 from argilla_server.schemas.v1.files import ObjectMetadata
-from fastapi import APIRouter, Body, Depends, HTTPException, UploadFile, Path, status, Security
+from fastapi import APIRouter, Body, Depends, File, HTTPException, UploadFile, Path, status, Security
 from fastapi.responses import StreamingResponse
 from minio import Minio
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -17,7 +17,7 @@ from argilla_server.security import auth
 from argilla_server.policies import DocumentPolicy, authorize, is_authorized
 from argilla_server.models import User
 from argilla_server.contexts import accounts, datasets, files
-from argilla_server.schemas.v1.documents import DocumentCreate, DocumentDelete, DocumentListItem
+from argilla_server.schemas.v1.documents import DocumentCreateRequest, DocumentDeleteRequest, DocumentListItem
 
 if TYPE_CHECKING:
     from argilla_server.models import Document
@@ -26,7 +26,7 @@ _LOGGER = logging.getLogger("documents")
 
 router = APIRouter(tags=["documents"])
 
-async def check_existing_document(db: AsyncSession, document_create: DocumentCreate):
+async def check_existing_document(db: AsyncSession, document_create: DocumentCreateRequest):
     # Add conditions for non-empty attributes
     conditions = []
     if document_create.pmid:
@@ -37,6 +37,8 @@ async def check_existing_document(db: AsyncSession, document_create: DocumentCre
         conditions.append(Document.doi == document_create.doi)
     if document_create.id:
         conditions.append(Document.id == document_create.id)
+    if document_create.reference:
+        conditions.append(Document.reference == document_create.reference)
 
     if not conditions:
         return None
@@ -58,7 +60,8 @@ async def check_existing_document(db: AsyncSession, document_create: DocumentCre
 @router.post("/documents", status_code=status.HTTP_201_CREATED, response_model=UUID)
 async def upload_document(
     *,
-    document_create: DocumentCreate,
+    document_create: DocumentCreateRequest = Depends(),
+    file_data: UploadFile | None = File(None),
     db: AsyncSession = Depends(get_async_db),
     client: Minio = Depends(files.get_minio_client),
     current_user: User = Security(auth.get_current_user)
@@ -72,17 +75,33 @@ async def upload_document(
             detail=f"Workspace with id `{document_create.workspace_id}` not found",
         )
     
-    object_path = files.get_pdf_s3_object_path(document_create.id)
-    existing_files = files.list_objects(client, workspace.name, prefix=object_path, include_version=False)
-    if not existing_files.objects and document_create.file_data is not None:
-        file_data_bytes = base64.b64decode(document_create.file_data)
-        response = files.put_object(
-            client, bucket=workspace.name, object=object_path, data=file_data_bytes, 
-            size=len(file_data_bytes), content_type="application/pdf", 
-            metadata=document_create.dict(include={"file_name": True, "pmid": True, "doi": True}))
+    if file_data is not None:
+        object_path = files.get_pdf_s3_object_path(document_create.id)
+        existing_files = files.list_objects(client, workspace.name, prefix=object_path, include_version=False, recursive=False)
+        # file_data_bytes = base64.b64decode(file_data)
+        file_data_bytes = await file_data.read()
+
+        put_object = False
+    
+        if existing_files.objects:
+            new_file_hash = files.compute_hash(file_data_bytes)
+            existing_hashes = [existing_file.etag.strip('"') for existing_file in existing_files.objects if existing_file.etag is not None]
+            
+            if new_file_hash not in existing_hashes:
+                put_object = True
+        else:
+            put_object = True
         
-        if not document_create.url:
+        if put_object:
+            
+            response = files.put_object(
+                client, bucket=workspace.name, object=object_path, data=file_data_bytes, 
+                size=len(file_data_bytes), content_type="application/pdf", 
+                metadata=document_create.dict(include={"file_name": True, "pmid": True, "doi": True}))
+            
             document_create.url = files.get_s3_object_url(response.bucket_name, response.object_name)
+            if file_data.filename and not document_create.file_name:
+                document_create.file_name = file_data.filename
     
     existing_document = await check_existing_document(db, document_create)
     if existing_document is not None:
@@ -159,10 +178,10 @@ async def get_document_by_id(
     return DocumentListItem.from_orm(document)
 
 
-@router.delete("/documents/workspace/{workspace_id}", status_code=status.HTTP_200_OK)
+@router.delete("/documents/workspace/{workspace_id}", status_code=status.HTTP_200_OK, response_model=int, description="Delete all documents by workspace_id, or a specific document by id, pmid, doi, or url")
 async def delete_documents_by_workspace_id(*,
     workspace_id: Union[UUID, str],
-    document_delete: DocumentDelete = Body(None),
+    document_delete: DocumentDeleteRequest = Body(None),
     db: AsyncSession = Depends(get_async_db),
     client: Minio = Depends(files.get_minio_client),
     current_user: User = Security(auth.get_current_user)